requirements.txt
inky==0.0.5
Pillow==4.2.1
matplotlib==2.0.2
mpl-finance==0.10.0
